move stat item row component to file
